Use with module from npm to handle global variable detection
This is a much better match for jade because it means we pick up the
globals at the time that the function is called, rather than when it is
first defined.

It’s also nice to leave the heavy lifting to the `with` module rather
than trying to do it ourselves.